Refactor: Remove sub widgets to a different file
